+++ macros/monsterFeatures/eyeRay.js
@@ -1,0 +1,1 @@
+await MidiQOL.completeItemUse(actor.items.find(i=>i.name === "Eye Rays"));